pyproject.toml
[project]
name = "ape"
version = "1.0.0"
description = "AI Productivity Engine - All-in-one AI platform"
authors = [{name = "APE Team"}]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    
    # Database
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.1",
    "psycopg2-binary>=2.9.9",
    "asyncpg>=0.29.0",
    
    # Redis
    "redis>=5.0.1",
    "hiredis>=2.3.2",
    
    # LLM & AI
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langgraph>=0.0.40",
    "langchain-groq>=0.0.3",
    "langchain-aws>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.20",
    "langchain-chroma>=0.1.0",
    "chromadb>=0.4.22",
    
    # AWS Services
    "boto3>=1.34.27",
    "botocore>=1.34.27",
    
    # External Services
    "firecrawl-py>=0.0.5",
    "requests>=2.31.0",
    "httpx>=0.26.0",
    
    # Authentication & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
    
    # Data Processing
    "pandas>=2.1.4",
    "openpyxl>=3.1.2",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.1.0",
    
    # Utilities
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "python-dateutil>=2.8.2",
    "pytz>=2024.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.26.0",
    "ruff>=0.1.14",
    "mypy>=1.8.0",
    "black>=24.1.1",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --strict-markers --cov=src"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/migrations/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
